Move StudentFormDialog logic from StudentDatabasePage
