+++ src/components/StudentForm/index.ts
@@ -1,2 +1,3 @@
 export * from "./FormSections";
 export * from "./StudentForm";
+export * from "./StudentFormDialog";
