+++ src/hooks/useStores.ts
@@ -11,30 +11,28 @@ interface StudentState {
   students: Student[];
 }
 
-export const useStudentStore = create<StudentState>()(
-  subscribeWithSelector((set) => {
-    return {
-      filter: [],
-      selectedStudent: null,
-      setFilter: (filter: StudentState["filter"]) => {
-        return set(() => {
-          return { filter };
-        });
-      },
-      setSelectedStudent: (selectedStudent: StudentState["selectedStudent"]) => {
-        return set(() => {
-          return { selectedStudent };
-        });
-      },
-      setStudents: (students: StudentState["students"]) => {
-        return set(() => {
-          return { students };
-        });
-      },
-      students: [],
-    };
-  }),
-);
+export const useStudentStore = create<StudentState>()((set) => {
+  return {
+    filter: [],
+    selectedStudent: null,
+    setFilter: (filter: StudentState["filter"]) => {
+      return set(() => {
+        return { filter };
+      });
+    },
+    setSelectedStudent: (selectedStudent: StudentState["selectedStudent"]) => {
+      return set(() => {
+        return { selectedStudent };
+      });
+    },
+    setStudents: (students: StudentState["students"]) => {
+      return set(() => {
+        return { students };
+      });
+    },
+    students: [],
+  };
+});
 
 interface WaitingListState {
   filter: FilterValue<WaitingListEntry>[];
@@ -45,30 +43,28 @@ interface WaitingListState {
   waitingList: WaitingListEntry[];
 }
 
-export const useWaitingListStore = create<WaitingListState>()(
-  subscribeWithSelector((set) => {
-    return {
-      filter: [],
-      selectedWaitingListEntry: null,
-      setFilter: (filter: WaitingListState["filter"]) => {
-        return set(() => {
-          return { filter };
-        });
-      },
-      setSelectedWaitingListEntry: (selectedWaitingListEntry: WaitingListState["selectedWaitingListEntry"]) => {
-        return set(() => {
-          return { selectedWaitingListEntry };
-        });
-      },
-      setWaitingList: (waitingList: WaitingListState["waitingList"]) => {
-        return set(() => {
-          return { waitingList };
-        });
-      },
-      waitingList: [],
-    };
-  }),
-);
+export const useWaitingListStore = create<WaitingListState>()((set) => {
+  return {
+    filter: [],
+    selectedWaitingListEntry: null,
+    setFilter: (filter: WaitingListState["filter"]) => {
+      return set(() => {
+        return { filter };
+      });
+    },
+    setSelectedWaitingListEntry: (selectedWaitingListEntry: WaitingListState["selectedWaitingListEntry"]) => {
+      return set(() => {
+        return { selectedWaitingListEntry };
+      });
+    },
+    setWaitingList: (waitingList: WaitingListState["waitingList"]) => {
+      return set(() => {
+        return { waitingList };
+      });
+    },
+    waitingList: [],
+  };
+});
 
 interface AppState {
   loading: boolean;
@@ -77,25 +73,23 @@ interface AppState {
   setRole: (role: AppState["role"]) => void;
 }
 
-export const useAppStore = create<AppState>()(
-  subscribeWithSelector((set) => {
-    return {
-      loading: true,
-      role: "staff",
+export const useAppStore = create<AppState>()((set) => {
+  return {
+    loading: true,
+    role: "staff",
 
-      setLoading: (loading: AppState["loading"]) => {
-        return set(() => {
-          return { loading };
-        });
-      },
-      setRole: (role: AppState["role"]) => {
-        return set(() => {
-          return { role };
-        });
-      },
-    };
-  }),
-);
+    setLoading: (loading: AppState["loading"]) => {
+      return set(() => {
+        return { loading };
+      });
+    },
+    setRole: (role: AppState["role"]) => {
+      return set(() => {
+        return { role };
+      });
+    },
+  };
+});
 
 interface FinalGradeReportState {
   setShouldDownload: (shouldDownload: FinalGradeReportState["shouldDownload"]) => void;
@@ -114,3 +108,19 @@ export const useFinalGradeReportStore = create<FinalGradeReportState>()(
     };
   }),
 );
+
+interface StudentFormState {
+  open: boolean;
+  setOpen: (open: StudentFormState["open"]) => void;
+}
+
+export const useStudentFormStore = create<StudentFormState>()((set) => {
+  return {
+    open: false,
+    setOpen: (open: StudentFormState["open"]) => {
+      return set(() => {
+        return { open };
+      });
+    },
+  };
+});
